Version 1.03: refactoring to mvp: - add BasePresenter interface - add BaseView interface - add MainContract interface - add MainPresenter - add ModelRepository
