+++ app/src/main/AndroidManifest.xml
@@ -12,15 +12,15 @@
         android:theme="@style/Theme.TrumpetABC"
         tools:targetApi="31">
         <activity
-            android:name=".NotesCardActivity"
+            android:name=".views.NotesCardActivity"
             android:exported="false"
             android:screenOrientation="portrait"/>
         <activity
-            android:name=".TonesActivity"
+            android:name=".views.TonesActivity"
             android:exported="false"
             android:screenOrientation="landscape" />
         <activity
-            android:name=".FingeringsActivity"
+            android:name=".views.FingeringsActivity"
             android:exported="false"
             android:screenOrientation="landscape" />
         <activity
