+++ app/src/main/java/com/dekram/trumpetabc/BasePresenter.kt
@@ -1,0 +1,6 @@
+package com.dekram.trumpetabc
+
+interface BasePresenter {
+
+    fun onDestroy()
+}
