+++ app/src/main/java/com/dekram/trumpetabc/BaseView.kt
@@ -1,0 +1,6 @@
+package com.dekram.trumpetabc
+
+interface BaseView<T> {
+
+    fun setPresenter(presenter: T)
+}
