+++ app/src/main/java/com/dekram/trumpetabc/MainContract.kt
@@ -1,0 +1,17 @@
+package com.dekram.trumpetabc
+
+interface MainContract {
+
+    interface Presenter: BasePresenter {
+        fun onViewCreated()
+    }
+
+    interface View: BaseView<Presenter> {
+
+        fun displayFingerings()
+
+        fun displayTones()
+
+        fun displayNotesCards()
+    }
+}
