+++ app/src/main/java/com/dekram/trumpetabc/model/ModelRepository.kt
@@ -1,0 +1,10 @@
+package com.dekram.trumpetabc.model
+
+interface ModelRepository {
+
+    fun loadFingerings()
+
+    fun loadTones()
+
+    fun loadNotesCards()
+}
