+++ app/src/main/java/com/dekram/trumpetabc/presenter/MainPresenter.kt
@@ -1,0 +1,15 @@
+package com.dekram.trumpetabc.presenter
+
+import com.dekram.trumpetabc.MainContract
+
+class MainPresenter(view: MainContract.View, dependencyInjector: DependencyInjector) :
+    MainContract.Presenter {
+
+    override fun onDestroy() {
+        TODO("Not yet implemented")
+    }
+
+    override fun onViewCreated() {
+        TODO("Not yet implemented")
+    }
+}
