+++ app/src/main/java/com/dekram/trumpetabc/views/NotesCardActivity.kt
@@ -1,4 +1,4 @@
-package com.dekram.trumpetabc
+package com.dekram.trumpetabc.views
 
 import android.content.Context
 import android.content.Intent
@@ -6,6 +6,7 @@ import android.media.MediaPlayer
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
+import com.dekram.trumpetabc.R
 import android.view.ViewGroup
 import android.widget.ImageView
 import android.widget.TextView
