+++ app/src/main/res/layout/activity_fingerings.xml
@@ -6,7 +6,7 @@
     android:layout_height="match_parent"
     android:background="@color/green_lime"
     android:padding="32dp"
-    tools:context=".FingeringsActivity">
+    tools:context=".views.FingeringsActivity">
 
     <androidx.cardview.widget.CardView
         android:layout_width="wrap_content"
