+++ app/src/main/res/layout/activity_notes_card.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context=".NotesCardActivity"
+    tools:context=".views.NotesCardActivity"
     android:background="@color/green_lime">
 
     <com.google.android.material.tabs.TabLayout
