+++ app/src/main/res/layout/activity_tones.xml
@@ -7,7 +7,7 @@
     android:background="@color/green_lime"
     android:orientation="horizontal"
     android:padding="32dp"
-    tools:context=".TonesActivity">
+    tools:context=".views.TonesActivity">
 
     <androidx.cardview.widget.CardView
         android:layout_width="wrap_content"
